requirements.txt
asgiref==3.7.2
Django==5.0
Pillow==10.1.0
psycopg2==2.9.9
python-decouple==3.8
sqlparse==0.4.4
tzdata==2023.3
djangorestframework